+++ src/components/DiagramTable/DiagramTable.jsx
@@ -26,6 +26,10 @@ import {
   formatStringWithSpaces,
   MakeDecimalPlaces,
 } from 'utilities/formatUtils';
+import {
+  sumAmountByCategory,
+  sumMonthlyIncomeAndExpenses,
+} from 'utilities/sumMonthlyTotals';
 
 const months = [
   { id: 1, name: 'January' },
@@ -99,47 +103,10 @@ const DiagramTableBase = () => {
     }
   };
 
-  const sumAmountByCategory = monthlyTotals => {
-    const categorySums = [];
-
-    for (const entry of monthlyTotals) {
-      const category = entry.category;
-      const amount = entry.amount;
-
-      const categoryObject = categorySums.find(
-        obj => obj.category === category
-      );
-
-      categoryObject
-        ? (categoryObject.amount += amount)
-        : categorySums.push({ category, amount });
-    }
-
-    return categorySums.filter(obj => obj.category !== 'income');
-  };
-
   const monthlyTotalsByCategory =
     Object.keys(monthlyTotals).length === 0
       ? null
       : sumAmountByCategory(monthlyTotals);
-
-  const sumMonthlyIncomeAndExpenses = monthlyTotals => {
-    let monthlyIncome = 0;
-    let monthlyExpenses = 0;
-
-    for (const entry of monthlyTotals) {
-      const isIncome = entry.isIncome;
-      const amount = entry.amount;
-
-      if (isIncome) {
-        monthlyIncome += amount;
-      } else {
-        monthlyExpenses += amount;
-      }
-    }
-
-    return { monthlyIncome, monthlyExpenses };
-  };
 
   const monthlyIncomeAndExpenses =
     Object.keys(monthlyTotals).length === 0
